# Conflicts:
#	src/main/java/de/uniks/stpmon/team_m/controller/IngameController.java
#	src/main/resources/de/uniks/stpmon/team_m/styles.css

+++ src/main/java/de/uniks/stpmon/team_m/dto/Chunk.java
@@ -3,7 +3,7 @@ package de.uniks.stpmon.team_m.dto;
 import java.util.List;
 
 public record Chunk(
-        List<Integer> data,
+        List<Long> data,
         int height,
         int width,
         int x,
